[scheduler] review comments
