+++ components/ee/ws-scheduler/pkg/scheduler/strategy.go
@@ -9,6 +9,8 @@ import (
 	"sort"
 	"time"
 
+	wsk8s "github.com/gitpod-io/gitpod/common-go/kubernetes"
+
 	log "github.com/sirupsen/logrus"
 	"golang.org/x/xerrors"
 
@@ -173,7 +175,7 @@ func (s *DensityAndExperience) Select(state *State, pod *corev1.Pod) (string, er
 	// Now we're only concerned with the user experience - but only for non-headless workspaces!
 	// Headless workspaces just mustn't break the user experience of non-headless ones. Hence we
 	// want to schedule headless workspaces as far away from regular ones as we can.
-	if isHeadlessWorkspace(pod) {
+	if wsk8s.IsHeadlessWorkspace(pod) {
 		// We try and find the least utilised node that still has regular workspaces on it.
 		// This way we place the headless workspace "on the other end" of the cluster, but
 		// don't prevent scale-down.
@@ -246,7 +248,7 @@ func fitsOnNode(pod *corev1.Pod, node *Node) bool {
 func freshWorkspaceCount(state *State, node *Node, freshSeconds int) int {
 	var count int
 	for _, p := range node.Pods {
-		if !isWorkspace(p) {
+		if !wsk8s.IsWorkspace(p) {
 			continue
 		}
 		if time.Since(p.ObjectMeta.CreationTimestamp.Time).Seconds() < float64(freshSeconds) {
@@ -300,10 +302,10 @@ func classifyNode(state *State, node *Node) int {
 func regularWorkspaceCount(state *State, node *Node) int {
 	var count int
 	for _, p := range node.Pods {
-		if !isWorkspace(p) {
+		if !wsk8s.IsWorkspace(p) {
 			continue
 		}
-		if isHeadlessWorkspace(p) {
+		if wsk8s.IsHeadlessWorkspace(p) {
 			continue
 		}
